Move more dackbox db inits behind feature flag
