+++ central/cve/service/singleton.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	legacyImageCVEDataStore "github.com/stackrox/rox/central/cve/datastore"
-	"github.com/stackrox/rox/central/globaldb/dackbox"
 	vulReqMgr "github.com/stackrox/rox/central/vulnerabilityrequest/manager/requestmgr"
 	"github.com/stackrox/rox/pkg/sync"
 )
@@ -15,7 +14,7 @@ var (
 
 func initialize() {
 	// This service is not used in postgres.
-	as = New(legacyImageCVEDataStore.Singleton(), dackbox.GetIndexQueue(), vulReqMgr.Singleton())
+	as = New(legacyImageCVEDataStore.Singleton(), vulReqMgr.Singleton())
 }
 
 // Singleton provides the instance of the Service interface to register.
