+++ central/deployment/datastore/singleton.go
@@ -1,10 +1,11 @@
 package datastore
 
 import (
+	"github.com/blevesearch/bleve"
 	"github.com/stackrox/rox/central/deployment/cache"
 	"github.com/stackrox/rox/central/deployment/datastore/internal/processtagsstore"
 	"github.com/stackrox/rox/central/globaldb"
-	"github.com/stackrox/rox/central/globaldb/dackbox"
+	globalDackBox "github.com/stackrox/rox/central/globaldb/dackbox"
 	"github.com/stackrox/rox/central/globalindex"
 	imageDatastore "github.com/stackrox/rox/central/image/datastore"
 	nfDS "github.com/stackrox/rox/central/networkgraph/flow/datastore"
@@ -12,6 +13,9 @@ import (
 	"github.com/stackrox/rox/central/processindicator/filter"
 	"github.com/stackrox/rox/central/ranking"
 	riskDS "github.com/stackrox/rox/central/risk/datastore"
+	"github.com/stackrox/rox/pkg/concurrency"
+	"github.com/stackrox/rox/pkg/dackbox"
+	"github.com/stackrox/rox/pkg/features"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/sync"
 )
@@ -25,11 +29,21 @@ var (
 )
 
 func initialize() {
-	ad = New(dackbox.GetGlobalDackBox(),
-		dackbox.GetKeyFence(),
+	var dackBox *dackbox.DackBox
+	var keyFence concurrency.KeyFence
+	var bleveIndex, processIndex bleve.Index
+	if !features.PostgresDatastore.Enabled() {
+		dackBox = globalDackBox.GetGlobalDackBox()
+		keyFence = globalDackBox.GetKeyFence()
+		bleveIndex = globalindex.GetGlobalIndex()
+		processIndex = globalindex.GetProcessIndex()
+	}
+	ad = New(dackBox,
+		keyFence,
+		// Process Tag store will be removed in 72.0
 		processtagsstore.New(globaldb.GetGlobalDB()),
-		globalindex.GetGlobalIndex(),
-		globalindex.GetProcessIndex(),
+		bleveIndex,
+		processIndex,
 		imageDatastore.Singleton(),
 		pbDS.Singleton(),
 		nfDS.Singleton(),
